condition show page testing

+++ db/migrate/20171204214948_create_conditions.rb
@@ -6,9 +6,9 @@ class CreateConditions < ActiveRecord::Migration[5.1]
       t.float :mean_temperature_f
       t.float :min_temperature_f
       t.float :mean_humidity
-      t.float :mean_visibility_miles
-      t.float :mean_wind_speed_mph
-      t.float :precipitation_inches
+      t.int :mean_visibility_miles
+      t.int :mean_wind_speed_mph
+      t.int :precipitation_inches
     end
   end
 end
